refactor(agr): restructure module
