+++ algorithm/src/main/kotlin/io/github/wliamp/agr/AutoConfig.kt
@@ -1,12 +1,11 @@
-package io.github.wliamp.agr.config
+package io.github.wliamp.agr
 
-import io.github.wliamp.agr.data.ReactiveQueue
 import org.springframework.boot.autoconfigure.AutoConfiguration
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 
 @AutoConfiguration
-internal class AlgorithmAutoConfig {
+internal class AutoConfig {
     @Bean
     @ConditionalOnMissingBean
     fun q(): ReactiveQueue<Any> = ReactiveQueue()
